+++ src/lib/kotlin/slatekit-cache/src/main/kotlin/slatekit/cache/CacheEntry.kt
@@ -16,7 +16,7 @@ package slatekit.cache
 import kotlinx.coroutines.runBlocking
 import slatekit.results.Outcome
 import slatekit.results.builders.Outcomes
-import slatekit.tracking.*
+import slatekit.telemetry.*
 import java.util.concurrent.atomic.AtomicReference
 
 /**
