+++ src/lib/kotlin/slatekit-telemetry/.editorconfig
