+++ src/lib/kotlin/slatekit-telemetry/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'slatekit-telemetry'
