+++ src/lib/kotlin/slatekit-telemetry/src/main/kotlin/slatekit/telemetry/Events.kt
@@ -1,4 +1,4 @@
-package slatekit.tracking
+package slatekit.telemetry
 
 import slatekit.utils.events.Event
 import slatekit.results.*
