+++ src/lib/kotlin/slatekit-telemetry/src/main/kotlin/slatekit/telemetry/Ratio.kt
@@ -1,4 +1,4 @@
-package slatekit.tracking
+package slatekit.telemetry
 
 data class Ratio(val name:String, val count:Long, val total:Long) {
     val value : Double = if(total == 0L) 0.0 else count / total.toDouble()
