+++ src/lib/kotlin/slatekit-telemetry/src/main/kotlin/slatekit/telemetry/Tag.kt
@@ -1,4 +1,4 @@
-package slatekit.tracking
+package slatekit.telemetry
 
 /**
  * Represents a single tag ( name/value pair ) to associate metrics with
