+++ src/lib/kotlin/slatekit-telemetry/src/main/kotlin/slatekit/telemetry/Tracked.kt
@@ -1,4 +1,4 @@
-package slatekit.tracking
+package slatekit.telemetry
 
 import slatekit.common.DateTime
