--- src/lib/kotlin/slatekit-tracking/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'slatekit-tracking'
