+++ src/lib/kotlin/slatekit/src/main/resources/templates/slatekit/api/files/build.txt
@@ -55,7 +55,7 @@ dependencies {
     compile "com.slatekit:slatekit-common:$slatekit_version"
     compile "com.slatekit:slatekit-context:$slatekit_version"
     compile "com.slatekit:slatekit-app:$slatekit_version"
-    compile "com.slatekit:slatekit-tracking:$slatekit_version"
+    compile "com.slatekit:slatekit-telemetry:$slatekit_version"
     compile "com.slatekit:slatekit-apis:$slatekit_version"
     compile "com.slatekit:slatekit-server:$slatekit_version"
     compile "com.slatekit:slatekit-serialization:$slatekit_version"
